requirements.txt
# Core dependencies
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.31
pydantic==1.10.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9

# File handling
pillow==10.4.0
aiofiles==23.1.0

# Utilities
python-dotenv==1.0.1
email-validator==2.0.0

# Production dependencies
gunicorn==21.2.0
httptools==0.6.0
watchgod==0.8.2

# Monitoring (optional)
prometheus-client==0.18.0

# Additional dependencies
alembic==1.13.2
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
requests==2.32.3
